backend/app/requirement.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
pydantic-settings
slowapi
pytest
pytest-asyncio
httpx
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
alembic
aiosqlite
greenlet
